arrow_csv::reader

Struct Format

pub struct Format {
    header: bool,
    delimiter: Option<u8>,
    escape: Option<u8>,
    quote: Option<u8>,
    terminator: Option<u8>,
    comment: Option<u8>,
    null_regex: NullRegex,
    truncated_rows: bool,
}

The format specification for the CSV file

Fields
header: booldelimiter: Option<u8>escape: Option<u8>quote: Option<u8>terminator: Option<u8>comment: Option<u8>null_regex: NullRegextruncated_rows: bool
Implementations

impl Format

pub fn with_header(self, has_header: bool) -> Self

Specify whether the CSV file has a header, defaults to false

When true, the first row of the CSV file is treated as a header row

pub fn with_delimiter(self, delimiter: u8) -> Self

Specify a custom delimiter character, defaults to comma ','

pub fn with_escape(self, escape: u8) -> Self

Specify an escape character, defaults to None

pub fn with_quote(self, quote: u8) -> Self

Specify a custom quote character, defaults to double quote '"'

pub fn with_terminator(self, terminator: u8) -> Self

Specify a custom terminator character, defaults to CRLF

pub fn with_comment(self, comment: u8) -> Self

Specify a comment character, defaults to None

Lines starting with this character will be ignored

pub fn with_null_regex(self, null_regex: Regex) -> Self

Provide a regex to match null values, defaults to ^$

pub fn with_truncated_rows(self, allow: bool) -> Self

Whether to allow truncated rows when parsing.

By default this is set to false and will error if the CSV rows have different lengths. When set to true then it will allow records with less than the expected number of columns and fill the missing columns with nulls. If the recordโs schema is not nullable, then it will still return an error.

pub fn infer_schema<R: Read>( &self, reader: R, max_records: Option<usize>, ) -> Result<(Schema, usize), ArrowError>

Infer schema of CSV records from the provided reader

If max_records is None, all records will be read, otherwise up to max_records records are read to infer the schema

Returns inferred schema and number of records read

fn build_reader<R: Read>(&self, reader: R) -> Reader<R>

Build a [csv::Reader] for this Format

fn build_parser(&self) -> Reader

Build a [csv_core::Reader] for this Format