Module arrow_csv::reader

CSV Reader

This CSV reader allows CSV files to be read into the Arrow memory model. Records are loaded in batches and are then converted from row-based data to columnar data.

Example:

let schema = Schema::new(vec![
    Field::new("city", DataType::Utf8, false),
    Field::new("lat", DataType::Float64, false),
    Field::new("lng", DataType::Float64, false),
]);

let file = File::open("test/data/uk_cities.csv").unwrap();

let mut csv = Reader::new(file, Arc::new(schema), false, None, 1024, None, None, None);
let batch = csv.next().unwrap().unwrap();

Structs

BufReader

CSV file reader

Decoder

A push-based interface for decoding CSV data from an arbitrary byte stream

ReaderBuilder

CSV file reader builder

Functions

infer_file_schema

Infer the schema of a CSV file by reading through the first n records of the file, with max_read_records controlling the maximum number of records to read.

infer_reader_schema

Infer schema of CSV records provided by struct that implements Read trait.

infer_schema_from_files

Infer schema from a list of CSV files by reading through first n records with max_read_records controlling the maximum number of records to read.

Type Definitions

Reader

CSV file reader using std::io::BufReader